+++ src/SyntaxParser.cc
@@ -199,12 +199,12 @@ namespace grumpy
             return a->GetParseIndex() < b->GetParseIndex();
         });
 
-        for (int i = 0; i < nodesVector.size(); ++i)
+        for (size_t i = 0; i < nodesVector.size(); ++i)
         {
             nodesVector[i]->Hide();
         }
 
-        for (int i = 0; i < nodesVector.size(); ++i)
+        for (size_t i = 0; i < nodesVector.size(); ++i)
         {
             nodesVector[i]->SetNodeColor(Util::InterpolateRainbow(i / static_cast<float>(nodesVector.size()), 0.25f));
         }
